requirements.txt
ultralytics
opencv-python
streamlit
torch
torchvision
Pillow
